+++ cli/src/pcluster/config/cluster_config.py
@@ -837,20 +837,14 @@ class Dashboards(Resource):
 class Monitoring(Resource):
     """Represent the Monitoring configuration."""
 
-    def __init__(
-        self, enable_detailed_monitoring: bool = None, logs: Logs = None, dashboards: Dashboards = None, **kwargs
-    ):
+    def __init__(self, detailed_monitoring: bool = None, logs: Logs = None, dashboards: Dashboards = None, **kwargs):
         super().__init__(**kwargs)
-        self.enable_detailed_monitoring = Resource.init_param(
-            enable_detailed_monitoring, default=DETAILED_MONITORING_ENABLED_DEFAULT
-        )
+        self.detailed_monitoring = Resource.init_param(detailed_monitoring, default=DETAILED_MONITORING_ENABLED_DEFAULT)
         self.logs = logs or Logs(implied=True)
         self.dashboards = dashboards or Dashboards(implied=True)
 
     def _register_validators(self, context: ValidatorContext = None):  # noqa: D102 #pylint: disable=unused-argument
-        self._register_validator(
-            DetailedMonitoringValidator, is_detailed_monitoring_enabled=self.enable_detailed_monitoring
-        )
+        self._register_validator(DetailedMonitoringValidator, is_detailed_monitoring_enabled=self.detailed_monitoring)
 
 
 # ---------------------- Others ---------------------- #
@@ -1708,7 +1702,7 @@ class BaseClusterConfig(Resource):
     @property
     def is_detailed_monitoring_enabled(self):
         """Return True if Detailed Monitoring is enabled."""
-        return self.monitoring.enable_detailed_monitoring
+        return self.monitoring.detailed_monitoring
 
     @property
     def is_dcv_enabled(self):